pyproject.toml
[tool.poetry]
name = "ilab"
version = "0.1.0"
description = ""
authors = ["daveb <davobain@yahoo.com.au>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.2.2"
numpy = "<2.0.0"
jupyterlab = "^4.2.4"
scikit-learn = "^1.5.1"
matplotlib = "^3.9.2"
seaborn = "^0.13.2"
requests = "^2.32.3"
boto3 = "^1.34.162"
pillow = "^10.4.0"
opencv-python = "^4.10.0.84"
keras = "^3.5.0"

tensorflow = "^2.16.2"
tensorflow-metal = "^1.0.0"

torch = "^2.4.0"
torchvision = "^0.19.0"
torchsummary = "^1.5.1"
aws = "^0.2.5"


[tool.poetry.group.dev.dependencies]
jupyter = "^1.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
